+++ src/commands/lint.rs
@@ -51,12 +51,13 @@ pub fn determine_lints(
 
     // determine the lints for the stacks
     for detected_stack in detected_stacks {
-        let stack_config = config.stack_config(detected_stack.stack.stack_type());
+        let stack_type = detected_stack.stack.stack_type();
+        let stack_config = config.stack_config(stack_type);
         // schedule either the override lints or the default lints
         let stack_lints = stack_config.and_then(|sc| sc.lint.as_ref());
         if let Some(overrides) = stack_lints.and_then(|lint| lint.replace.as_ref()) {
             for override_lint in overrides {
-                let executable = conc::Executable::from(override_lint);
+                let executable = override_lint.to_executable(Operation::Lint, stack_type);
                 result.push(conc::Runnable::Single(executable));
             }
         } else {
@@ -71,7 +72,7 @@ pub fn determine_lints(
         }
         if let Some(additions) = stack_lints.and_then(|lint| lint.add.as_ref()) {
             for addition in additions {
-                let executable = conc::Executable::from(addition);
+                let executable = addition.to_executable(Operation::Lint, stack_type);
                 result.push(conc::Runnable::Single(executable));
             }
         }